code cleanup & change /game/save
