+++ controllers/leaderboardController.js
@@ -17,3 +17,14 @@ export const updatePokemonScore = async (req, res) => {
     res.status(500).send(error.message);
   }
 };
+
+export const getLeaderboard = async (req, res) => {
+  try {
+    const leaderboardData = await Pokemon.find().sort({ score: -1 }).limit(10);
+
+    res.json(leaderboardData);
+  } catch (error) {
+    console.error("Error fetching leaderboard data", error);
+    res.status(500).send("Error fetching leaderboard data");
+  }
+};
