+++ controllers/pokemonController.js
@@ -61,14 +61,3 @@ export const pokemonById = async (req, res) => {
     res.status(500).send("Error fetching Pokémon data");
   }
 };
-
-export const getLeaderboard = async (req, res) => {
-  try {
-    const leaderboardData = await Pokemon.find().sort({ score: -1 }).limit(10);
-
-    res.json(leaderboardData);
-  } catch (error) {
-    console.error("Error fetching leaderboard data", error);
-    res.status(500).send("Error fetching leaderboard data");
-  }
-};
